move internal/http/app/hello to internal/api/hello
