+++ hw12_13_14_15_calendar/internal/api/hello/app.go
@@ -1,4 +1,4 @@
-// http/app/hello - приложение с контроллером hello.
+// api/hello - приложение с контроллером hello.
 // Представляет собой пакет с обработчиками для http-сервера основанного на http/web/mux мультиплексоре.
 //
 // Основная задача - связь бизнес-логики business/hello, запросов через http сервер и формирование ответов.
